+++ stelvio/aws/cognito/__init__.py
@@ -1,0 +1,25 @@
+from stelvio.aws.cognito.identity_provider import IdentityProvider
+from stelvio.aws.cognito.types import (
+    PasswordPolicy,
+    PasswordPolicyDict,
+    TriggerConfigDict,
+    UserPoolClientConfig,
+    UserPoolClientConfigDict,
+    UserPoolConfig,
+    UserPoolConfigDict,
+)
+from stelvio.aws.cognito.user_pool import UserPool
+from stelvio.aws.cognito.user_pool_client import UserPoolClient
+
+__all__ = [
+    "IdentityProvider",
+    "PasswordPolicy",
+    "PasswordPolicyDict",
+    "TriggerConfigDict",
+    "UserPool",
+    "UserPoolClient",
+    "UserPoolClientConfig",
+    "UserPoolClientConfigDict",
+    "UserPoolConfig",
+    "UserPoolConfigDict",
+]