+++ tests/aws/test_keyword_only_constructor_params.py
@@ -7,6 +7,8 @@ from stelvio.aws.api_gateway.api import Api, _create_custom_domain
 from stelvio.aws.cloudfront.cloudfront import CloudFrontDistribution
 from stelvio.aws.cloudfront.origins.components.url import Url
 from stelvio.aws.cloudfront.router import Router
+from stelvio.aws.cognito.user_pool import UserPool
+from stelvio.aws.cognito.user_pool_client import UserPoolClient
 from stelvio.aws.cron import Cron
 from stelvio.aws.dynamo_db import DynamoSubscription, DynamoTable
 from stelvio.aws.email import Email
@@ -53,6 +55,9 @@ from stelvio.aws.topic import Topic, TopicQueueSubscription, TopicSubscription
         (TopicSubscription.__init__, "customize"),
         (Cron.__init__, "tags"),
         (Cron.__init__, "customize"),
+        (UserPool.__init__, "tags"),
+        (UserPool.__init__, "customize"),
+        (UserPoolClient.__init__, "customize"),
         (_create_custom_domain, "tags"),
     ],
 )